+++ src/components/Form/Teacher/FormTeacher.jsx
@@ -12,6 +12,7 @@ import ScheduleStep from "./Steps/ScheduleStep";
 import PricingStep from "./Steps/PricingStep";
 import { createTeacher } from '../../../services/teacher.services'
 import { uploadImage } from "../../../services/utils";
+import { processImageUpload } from "../../../helpers/processImageUpload";
 
 
 
@@ -35,43 +36,7 @@ const MultiStepForm = () => {
   };
 
   useEffect(() => {
-    const processImageUpload = async () => {
-      if (formData.profileImageUrl) {
-        const imageUrl = formData.profileImageUrl;
-  
-        // Verifica que imageUrl sea un string antes de usar startsWith
-        if (typeof imageUrl === 'string' && imageUrl.startsWith('data:image/')) {
-          console.log('La imagen es un data URL');
-          
-          // Obtener el tipo de la imagen del data URL
-          const mimeType = imageUrl.match(/data:(.*?);base64/)[1];
-    
-          // Convertir el data URL en un Blob
-          const byteString = atob(imageUrl.split(',')[1]);
-          const arrayBuffer = new Uint8Array(byteString.length);
-          for (let i = 0; i < byteString.length; i++) {
-            arrayBuffer[i] = byteString.charCodeAt(i);
-          }
-          const blob = new Blob([arrayBuffer], { type: mimeType });
-          const file = new File([blob], `profile-image.${mimeType.split('/')[1]}`, { type: mimeType });
-          
-          // Esto espera que uploadImage retorne una URL
-          const uploadedImageUrl = await uploadImage(file, mimeType);
-  
-          // Guarda la URL en formData.profileImageUrl
-          setFormData((prevFormData) => ({
-            ...prevFormData,
-            profileImageUrl: uploadedImageUrl
-          }));
-        } else {
-          console.log('La URL de la imagen no es un data URL o no es un string.');
-        }
-      } else {
-        console.log(JSON.stringify(formData));
-      }
-    };
-  
-    processImageUpload();
+    processImageUpload(formData, uploadImage, setFormData); // svg to s3
   }, [formData]);
